Move keycloak to server only
